bug fixing

+++ lib/actions/user.action.ts
@@ -21,7 +21,6 @@ export async function getUserById(params: any) {
     return user;
   } catch (error) {
     console.log(error);
-
     throw error;
   }
 }
@@ -31,10 +30,10 @@ export async function createUser(userData: CreateUserParams) {
     connectToDataBase();
 
     const newUser = await User.create(userData);
+
     return newUser;
   } catch (error) {
     console.log(error);
-
     throw error;
   }
 }
@@ -44,13 +43,14 @@ export async function updateUser(params: UpdateUserParams) {
     connectToDataBase();
 
     const { clerkId, updateData, path } = params;
+
     await User.findOneAndUpdate({ clerkId }, updateData, {
       new: true,
     });
+
     revalidatePath(path);
   } catch (error) {
     console.log(error);
-
     throw error;
   }
 }
@@ -60,31 +60,29 @@ export async function deleteUser(params: DeleteUserParams) {
     connectToDataBase();
 
     const { clerkId } = params;
+
     const user = await User.findOneAndDelete({ clerkId });
 
     if (!user) {
       throw new Error("User not found");
     }
 
-    // Delete every thing from database comment questions likes
+    // Delete user from database
+    // and questions, answers, comments, etc.
 
-    // get all  questions  ids;
-    // const userQuestionIds = await Question.find({
-    //   author: user._id,
-    // }).distinct("_id");
+    // get user question ids
+    // const userQuestionIds = await Question.find({ author: user._id}).distinct('_id');
 
-    // delete user question
-    await Question.deleteMany({
-      author: user._id,
-    });
+    // delete user questions
+    await Question.deleteMany({ author: user._id });
 
-    // Todo dlete user anser and and comments and questions
+    // TODO: delete user answers, comments, etc.
 
-    const deleteUser = await User.findByIdAndDelete(user._id);
-    return deleteUser;
+    const deletedUser = await User.findByIdAndDelete(user._id);
+
+    return deletedUser;
   } catch (error) {
     console.log(error);
-
     throw error;
   }
 }
